Proposal for patching SonarQube rule 2164

+++ widgets/compositetable/org.eclipse.nebula.widgets.compositetable/src/org/eclipse/nebula/widgets/compositetable/day/internal/DayEditorCalendarableItemControl.java
@@ -110,7 +110,7 @@ public class DayEditorCalendarableItemControl extends Canvas implements ICalenda
 	 */
 	private RGB lighten(RGB color, float amount) {
 		float[] hsb = color.getHSB();
-		float b = hsb[2] + hsb[2] * amount;
+		float b = hsb[2] + ((double) hsb[2] * (double) amount) ;
 		if (b < 0) b=0;
 		if (b > 1) b=1;
 		return new RGB(hsb[0], hsb[1], b);
