+++ widgets/compositetable/org.eclipse.nebula.widgets.compositetable/src/org/eclipse/nebula/widgets/compositetable/day/internal/TimeSlot.java
@@ -118,7 +118,7 @@ public class TimeSlot extends Canvas {
 	 */
 	private RGB lighten(RGB color, float amount) {
 		float[] hsb = color.getHSB();
-		float b = hsb[2] + hsb[2] * amount;
+		float b = hsb[2] + ((double) hsb[2] * (double) amount) ;
 		if (b < 0) b=0;
 		if (b > 1) b=1;
 		return new RGB(hsb[0], hsb[1], b);
