+++ widgets/compositetable/org.eclipse.nebula.widgets.compositetable/src/org/eclipse/nebula/widgets/compositetable/month/internal/Day.java
@@ -115,7 +115,7 @@ public class Day extends Canvas implements PaintListener, DisposeListener {
 	 */
 	private RGB lighten(RGB color, float amount) {
 		float[] hsb = color.getHSB();
-		float b = hsb[2] + hsb[2] * amount;
+		float b = hsb[2] + ((double) hsb[2] * (double) amount) ;
 		if (b < 0)
 			b = 0;
 		if (b > 1)
